+++ lib/src/lib/components/form-fields/button-toggle-field/button-toggle-field.component.ts
@@ -39,6 +39,7 @@ export class ButtonToggleFieldComponent extends FormComponent<FormFieldButtonTog
         : options?.buttonOptions.find(o => o.value === this.controlValue())?.label) ?? '-'
     );
   });
+  protected readonly hideSelection = computed(() => this._options()?.hideSelection ?? false);
 
   public constructor() {
     super();