+++ lib/src/lib/components/form-fields/button-toggle-field/button-toggle-field.model.ts
@@ -15,6 +15,7 @@ export interface FormFieldButtonToggleOptions extends FormFieldBaseOptions {
     };
   }[];
   color?: ThemePalette;
+  hideSelection?: boolean;
   deselectOnClick?: boolean;
 }